Sépare les classes en plusieurs fichiers
